requirements.txt
streamlit
requests
python-dotenv
openai
anthropic
psutil
